% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann.R
\name{add_missed_cleavage_column}
\alias{add_missed_cleavage_column}
\title{Add a column 'Missed.Cleavages' with integer number of missed cleavages}
\usage{
add_missed_cleavage_column(parquet)
}
\arguments{
\item{parquet}{A DIA-NN outputted .parquet report}
}
\value{
A parquet report with an added column 'Missed.Cleavages'
}
\description{
The calculation is based on internal R/K not suffixed by P. The calculation defaults to use the 'Stripped.Sequence' column of the .parquet report passed
}
\examples{
\dontrun{p <- add_missed_cleavage_column(parquet)}
}
