% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann.R
\name{extract_sample_ids}
\alias{extract_sample_ids}
\title{Extract sample ID and technical ID from a parquet report}
\usage{
extract_sample_ids(t, pattern)
}
\arguments{
\item{t}{A tibble with a DIA-NN parquet report}

\item{patter}{A regex pattern with exactly two groups, the first being the sample id and the second being the replicate id.}
}
\value{
An altered tibble which contains two new columns, Sample.ID and Replicate.ID
}
\description{
Apply a regex pattern containing two groups to the 'Run' column of a .parquet report
to extract a sample id and a replicate id.
}
\examples{
tibble \%>\% extract_sample_ids(., pattern="P167_R2011_X3961_[A-Z][0-9]+_S([A-Z])([0-9])")

}
