% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_phospho.R
\name{join_site_report}
\alias{join_site_report}
\title{Combine site-level and global parquet reports appropriately}
\usage{
join_site_report(
  site_parquet,
  full_parquet,
  join_by = c("Sample.ID", "Replicate.ID")
)
}
\arguments{
\item{site_parquet}{An object containing the site-level parquet report output by dia-nn.}

\item{full_parquet}{An object containing the full (report.parquet) output by dia-nn.}
}
\value{
A data frame containing the joined site-level report (by precursor lib index and auxillary columns)
}
\description{
Combine site-level and global parquet reports appropriately
}
