% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_presentable}
\alias{make_presentable}
\title{Make a figure presentable}
\usage{
make_presentable(
  base_size = 12,
  margin = 15,
  font = "Arial",
  legend.position = "top"
)
}
\description{
Make a figure presentable
}
