% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_presentable_table}
\alias{make_presentable_table}
\title{Make a gt table presentable}
\usage{
make_presentable_table(
  gt_tbl,
  font = "Computer Modern",
  title = NULL,
  subtitle = NULL,
  footnote = NULL,
  source_note = NULL,
  add_horizontal_rules = FALSE
)
}
\arguments{
\item{gt_tbl}{a gt table}

\item{font}{Typeface}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{footnote}{Plot footnote}

\item{source_note}{Plot sourcenote}

\item{add_horizontal_rules}{Separate rows by horizontal rules}
}
\description{
Style the gt-table in a latex booktabs-ish style.
}
