% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann.R
\name{parquet_summarize_ids_cleavage}
\alias{parquet_summarize_ids_cleavage}
\title{Summarize a parquet report into key summary statistics}
\usage{
parquet_summarize_ids_cleavage(parquet, group.id, replicate.id)
}
\arguments{
\item{parquet}{a DIA-NN output .parquet report}

\item{group.id}{The upper level grouping column (str) \link{Experimental group}}

\item{replicate.id}{The lower level grouping column (str) \link{Technical replicate ID}}
}
\value{
A summarized dataframe with columns: group.id, replicate.id, Protein.Groups, Precursors, Peptides, Missed.Cleavage.Rate
}
\description{
Summarize a parquet report into key summary statistics
}
