% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann.R
\name{summarise_to_peptides}
\alias{summarise_to_peptides}
\title{Summarize a DIA-NN parquet report to peptide level using MaxLFQ algorithm}
\usage{
summarise_to_peptides(t, sample.id, replicate.id)
}
\arguments{
\item{t}{A tibble with a DIA-NN parquet report}

\item{sample.id}{Name of column containing a sample id}

\item{replicate.id}{Name of column containing a replicate id}
}
\description{
Summarize a DIA-NN parquet report to peptide level using MaxLFQ algorithm
}
