% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann.R
\name{summarise_to_pgs}
\alias{summarise_to_pgs}
\title{Summarize parquet report to protein groups and their intensities}
\usage{
summarise_to_pgs(t, groups, log2.transform = TRUE)
}
\arguments{
\item{t}{A tibble with a DIA-NN output parquet report}

\item{groups}{A character vector of columns to group the proteins by}

\item{log2.transform}{Return protein intensities (i) as log2(i+1)}
}
\value{
A summarized report with proteins and their intensities grouped by the 'groups' parameter.
}
\description{
Summarize parquet report to protein groups and their intensities
}
\examples{
tibble \%>\% summarize_to_pogs(groups=c("Sample.ID", "Replicate.ID"))
}
