\name{valid8assess}
\alias{valid8assess}
\title{
checks an /assess/
}
\description{
This function checks /assess/ objects
}
\usage{
valid8assess(object)
}
\arguments{
\item{object}{The assess object to be validated.}
}
\details{
It is the validity method for /assess/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
