\name{valid8chain}
\alias{valid8chain}
\title{
checks a /chain/
}
\description{
This function checks /chain/ objects
}
\usage{
valid8chain(object)
}
\arguments{
\item{object}{The chain object to be validated.}
}
\details{
It is the validity method for /chain/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
