\name{valid8distri}
\alias{valid8distri}
\title{
checks a /distri/
}
\description{
This function checks /distri/ objects
}
\usage{
valid8distri(object)
}
\arguments{
\item{object}{The distri object to be validated.}
}
\details{
It is the validity method for /distri/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
