\name{valid8nod}
\alias{valid8nod}
\title{
checks a /nod/
}
\description{
This function checks /nod/ objects
}
\usage{
valid8nod(object)
}
\arguments{
\item{object}{The nod object to be validated.}
}
\details{
It is the validity method for /nod/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
