\name{valid8probab}
\alias{valid8probab}
\title{
checks a /probab/
}
\description{
This function checks /probab/ objects
}
\usage{
valid8probab(object)
}
\arguments{
\item{object}{The probab object to be validated.}
}
\details{
It is the validity method for /probab/ objects.
}
\value{
TRUE when the object seems acceptable else a character describing the
error(s)
}
\keyword{classes}
