% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{count_missed_cleavage_string}
\alias{count_missed_cleavage_string}
\title{Count number of missed clevages in string of amino acids/peptide sequence}
\usage{
count_missed_cleavage_string(seq)
}
\arguments{
\item{seq}{A string representation of amino acid sequence.}
}
\value{
An integer
}
\description{
Counts the number of internal R/K not suffixed by proline in an AA-string
}
\examples{
count_missed_cleavage_string("AAAALLLRE")
}
