% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann.R
\name{parquet_summarize_ids_cleavage}
\alias{parquet_summarize_ids_cleavage}
\title{Summarize a parquet report into key summary statistics}
\usage{
parquet_summarize_ids_cleavage(parquet, groups)
}
\arguments{
\item{parquet}{a DIA-NN output .parquet report}

\item{groups}{A vector of columns to group the summary by. Could be c("Sample.ID", "Replicate.ID")}
}
\value{
A summarized dataframe with columns: *groups, Protein.Groups, Precursors, Peptides, Missed.Cleavage.Rate
}
\description{
Summarize a parquet report into key summary statistics
}
